Significant improvement in Fudge encode/decode performance.
Turns out that when I moved to the new size computation system, the
added performance overhead of size computations were pretty significant.
By moving that logic to a base class and improving the no-taxonomy size
case, we regain the speed improvement over Java serialization.

Along the way I changed the way we worked with taxonomies on decoding
messages for more speed improvments in the common case as well, to limit
the amount of code we're working with when taxonomies aren't in use.