Full implementation of new message encoding specification.
This involves having a new FudgeMsgEnvelope, which is the primary object
that is encoded on the wire with the header, and then a series of fields,
terminated with the special field type.
There were a lot of small objects required here, in part due to the removal
of the taxonomy requirement for some sub-message processing in FudgeMsgFieldType,
which allowed that parameter to be removed from the overridden method.

The Encoder/Decoder aren't particularly pretty at the moment quite frankly.
They need refactoring once everybody's agreed to the encoding spec.